requirements.txt
annotated-types==0.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
elastic-transport==8.4.0
elasticsearch==7.13.0
idna==3.4
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
